Struct google_cloudscheduler1_beta1::Job

pub struct Job {
    pub status: Option<Status>,
    pub description: Option<String>,
    pub name: Option<String>,
    pub schedule: Option<String>,
    pub schedule_time: Option<String>,
    pub http_target: Option<HttpTarget>,
    pub app_engine_http_target: Option<AppEngineHttpTarget>,
    pub last_attempt_time: Option<String>,
    pub state: Option<String>,
    pub pubsub_target: Option<PubsubTarget>,
    pub time_zone: Option<String>,
    pub retry_config: Option<RetryConfig>,
    pub attempt_deadline: Option<String>,
    pub user_update_time: Option<String>,
}

Configuration for a job. The maximum allowed size for a job is 100KB.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

locations jobs pause projects (response)

locations jobs get projects (response)

locations jobs create projects (request|response)

locations jobs resume projects (response)

locations jobs run projects (response)

locations jobs patch projects (request|response)

Fields
status: Option<Status>

Output only. The response from the target for the last attempted execution.

description: Option<String>

Optionally caller-specified in CreateJob or UpdateJob.

A human-readable description for the job. This string must not contain more than 500 characters.

name: Option<String>

Optionally caller-specified in CreateJob, after which it becomes output only.

The job name. For example: projects/PROJECT_ID/locations/LOCATION_ID/jobs/JOB_ID.

PROJECT_ID can contain letters ([A-Za-z]), numbers ([0-9]), hyphens (-), colons (:), or periods (.). For more information, see Identifying projects

LOCATION_ID is the canonical ID for the job's location. The list of available locations can be obtained by calling ListLocations. For more information, see https://cloud.google.com/about/locations/.

JOB_ID can contain only letters ([A-Za-z]), numbers ([0-9]), hyphens (-), or underscores (_). The maximum length is 500 characters.

schedule: Option<String>

Required, except when used with UpdateJob.

Describes the schedule on which the job will be executed.

The schedule can be either of the following types:

Crontab

English-like schedule

As a general rule, execution n + 1 of a job will not begin until execution n has finished. Cloud Scheduler will never allow two simultaneously outstanding executions. For example, this implies that if the n+1th execution is scheduled to run at 16:00 but the nth execution takes until 16:15, the n+1th execution will not start until 16:15. A scheduled start time will be delayed if the previous execution has not ended when its scheduled time occurs.

If retry_count > 0 and a job attempt fails, the job will be tried a total of retry_count times, with exponential backoff, until the next scheduled start time.

schedule_time: Option<String>

Output only. The next time the job is scheduled. Note that this may be a retry of a previously failed attempt or the next execution time according to the schedule.

http_target: Option<HttpTarget>

HTTP target.

app_engine_http_target: Option<AppEngineHttpTarget>

App Engine HTTP target.

last_attempt_time: Option<String>

Output only. The time the last job attempt started.

state: Option<String>

Output only. State of the job.

pubsub_target: Option<PubsubTarget>

Pub/Sub target.

time_zone: Option<String>

Specifies the time zone to be used in interpreting schedule. The value of this field must be a time zone name from the tz database.

Note that some time zones include a provision for daylight savings time. The rules for daylight saving time are determined by the chosen tz. For UTC use the string "utc". If a time zone is not specified, the default will be in UTC (also known as GMT).

retry_config: Option<RetryConfig>

Settings that determine the retry behavior.

attempt_deadline: Option<String>

The deadline for job attempts. If the request handler does not respond by this deadline then the request is cancelled and the attempt is marked as a DEADLINE_EXCEEDED failure. The failed attempt can be viewed in execution logs. Cloud Scheduler will retry the job according to the RetryConfig.

The allowed duration for this deadline is:

For HTTP targets, between 15 seconds and 30 minutes.

For App Engine HTTP targets, between 15 seconds and 24 hours.

For PubSub targets, this field is ignored.

user_update_time: Option<String>

Output only. The creation time of the job.